Assets count endpoint splitted in four
