+++ iot_api/user_api/endpoints/__init__.py
@@ -1,0 +1,4 @@
+from iot_api.user_api.endpoints.endpoints import *
+from iot_api.user_api.endpoints.Inventory import (AssetsListAPI,
+    AssetsPerVendorCountAPI, AssetsPerGatewayCountAPI,
+    AssetsPerDatacollectorCountAPI, AssetsPerTagCountAPI)
